Half Way Done With 1.8.x

+++ versions/1.8.x/source/Craft.Net.Client/Events/HealthAndFoodEventArgs.cs
@@ -8,10 +8,10 @@ namespace Craft.Net.Client.Events
     public class HealthAndFoodEventArgs : EventArgs
     {
         public float OldHealth { get; set; }
-        public short OldFood { get; set; }
+        public int OldFood { get; set; }
         public float OldFoodSaturation { get; set; }
         public float Health { get; set; }
-        public short Food { get; set; }
+        public int Food { get; set; }
         public float FoodSaturation { get; set; }
 
         internal bool IsChanged()
@@ -19,7 +19,7 @@ namespace Craft.Net.Client.Events
             return OldHealth != Health || OldFood != Food || OldFoodSaturation != FoodSaturation;
         }
 
-        public HealthAndFoodEventArgs(float oldHealth, short oldFood, float oldFoodSaturation)
+        public HealthAndFoodEventArgs(float oldHealth, int oldFood, float oldFoodSaturation)
         {
             OldHealth = oldHealth;
             OldFood = oldFood;
